requirements.txt
accelerate
bitsandbytes
faiss-cpu
ipykernel
llama-cpp-python
langchain
langchain-anthropic
langchain-community
langchain-core
langchain-openai
langgraph
langgraph-checkpoint-sqlite
langgraph-cli[inmem]
langgraph-prebuilt
langgraph-sdk
langsmith
notebook
optimum
pillow
python-dotenv
sentence-transformers
streamlit
tavily-python
timm
torch
transformers
trustcall
wikipedia
PyMuPDF
pytesseract
langchain-text-splitters
chromadb
protobuf==3.20.3
rank_bm25
google-api-python-client
gradio